Prevent doing N+1 queries, to improve overall performance.

Depends on;
https://github.com/django-oscar/django-oscar/pull/4350 & https://github.com/django-oscar/django-oscar-odin/pull/39/files to be merged.

Dummy DB with barely any attributes;
`1000001 products successfully indexed in 364.46 seconds`

PRD DB with 384846 products and 8876914 total attribute values:
`Processed chunk 1 of 384 (1000/384846 products indexed) in 2.53 seconds`

Which would first be > 10 seconds for each chunk, sometimes even 20 seconds.